+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.3...HEAD)
 ### Changed
 - Writing empty index.php files into our cache directories, so the directory listing doesn't work there
+- CacheHelper file permissions for dummy index.php
 
 ## [v0.3](https://github.com/ppfeufer/eve-online-fitting-manager/releases/tag/v0.3) - 2017-08-08
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.2...v0.3)